backend/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.2.0

# Google OAuth
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Environment variables
python-dotenv==1.0.1

# Email validation
email-validator==2.2.0

# Pydantic settings
pydantic[email]==2.9.2
pydantic-settings==2.6.0
